requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2024.1
